Decompose Key Emulate (#616)
**Background**

Step for integrate infared emulate

**Changes**

* Start/Stop/Error/OpenApp Helpers

**Test plan**

* Green pipeline
* Try emulate keys
