+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 - [Feature] FapHub manifest support offline mode and add target support for buttons 
 - [Refactor] Migrate all feature components modules from KSP to Anvil
 - [Refactor] Migrate to Ktorfit
-- [Refactor] Migrate key emulate to new module
+- [Refactor] Migrate key emulate to new module, decompose Emulate Helper
 - [Refactor] Key Screen state in API and KeyCard with state
 
 # 1.5.1
