+++ components/keyemulate/impl/src/main/java/com/flipperdevices/keyemulate/helpers/StopEmulateHelper.kt
@@ -1,0 +1,45 @@
+package com.flipperdevices.keyemulate.helpers
+
+import com.flipperdevices.bridge.api.manager.FlipperRequestApi
+import com.flipperdevices.bridge.api.model.FlipperRequestPriority
+import com.flipperdevices.bridge.api.model.wrapToRequest
+import com.flipperdevices.core.di.AppGraph
+import com.flipperdevices.core.log.LogTagProvider
+import com.flipperdevices.core.log.info
+import com.flipperdevices.protobuf.app.appButtonReleaseRequest
+import com.flipperdevices.protobuf.app.appExitRequest
+import com.flipperdevices.protobuf.main
+import com.squareup.anvil.annotations.ContributesBinding
+import kotlinx.coroutines.flow.flowOf
+import javax.inject.Inject
+
+interface StopEmulateHelper {
+    suspend fun onStop(requestApi: FlipperRequestApi)
+}
+
+@ContributesBinding(AppGraph::class, StopEmulateHelper::class)
+class StopEmulateHelperImpl @Inject constructor() : StopEmulateHelper, LogTagProvider {
+    override val TAG = "StopEmulateHelper"
+
+    override suspend fun onStop(requestApi: FlipperRequestApi) {
+        info { "stopEmulateInternal" }
+
+        val appButtonResponse = requestApi.request(
+            flowOf(
+                main {
+                    appButtonReleaseRequest = appButtonReleaseRequest { }
+                }.wrapToRequest(FlipperRequestPriority.FOREGROUND)
+            )
+        )
+        info { "App button stop response: $appButtonResponse" }
+
+        val appExitResponse = requestApi.request(
+            flowOf(
+                main {
+                    appExitRequest = appExitRequest { }
+                }.wrapToRequest(FlipperRequestPriority.FOREGROUND)
+            )
+        )
+        info { "App exit response: $appExitResponse" }
+    }
+}
