+++ components/widget/impl/src/main/java/com/flipperdevices/widget/impl/di/WidgetComponent.kt
@@ -2,7 +2,6 @@ package com.flipperdevices.widget.impl.di
 
 import com.flipperdevices.bridge.dao.api.delegates.WidgetDataApi
 import com.flipperdevices.core.di.AppGraph
-import com.flipperdevices.keyemulate.api.EmulateHelper
 import com.flipperdevices.widget.impl.providers.BaseWidgetProvider
 import com.flipperdevices.widget.impl.storage.WidgetStateStorage
 import com.flipperdevices.widget.impl.tasks.StartEmulateWorker
@@ -16,7 +15,6 @@ import com.squareup.anvil.annotations.ContributesTo
 @ContributesTo(AppGraph::class)
 interface WidgetComponent {
     val widgetDataApi: WidgetDataApi
-    val emulateHelper: EmulateHelper
     val widgetStateStorage: WidgetStateStorage
     val invalidateWidgetsHelper: InvalidateWidgetsHelper
